+++ QSB/Events/EventList.cs
@@ -1,5 +1,9 @@
-﻿using QSB.ElevatorSync;
+﻿using QSB.Animation;
+using QSB.DeathSync;
+using QSB.ElevatorSync;
 using QSB.GeyserSync;
+using QSB.Tools;
+using QSB.TransformSync;
 
 namespace QSB.Events
 {
