+++ QSB/QSB.csproj
@@ -119,6 +119,8 @@
     <Compile Include="Animation\AnimControllerPatch.cs" />
     <Compile Include="Animation\AnimFloatParam.cs" />
     <Compile Include="Animation\AnimTrigger.cs" />
+    <Compile Include="Animation\AnimTriggerMessage.cs" />
+    <Compile Include="DeathSync\DeathPatches.cs" />
     <Compile Include="ElevatorSync\ElevatorDirection.cs" />
     <Compile Include="ElevatorSync\QSBElevator.cs" />
     <Compile Include="ElevatorSync\ElevatorEvent.cs" />
@@ -126,26 +128,25 @@
     <Compile Include="ElevatorSync\ElevatorMessage.cs" />
     <Compile Include="ElevatorSync\ElevatorPatches.cs" />
     <Compile Include="Events\EventNames.cs" />
-    <Compile Include="Events\PlayerDeathEvent.cs" />
-    <Compile Include="Events\PlayerFlashlightEvent.cs" />
+    <Compile Include="DeathSync\PlayerDeathEvent.cs" />
+    <Compile Include="Tools\PlayerFlashlightEvent.cs" />
     <Compile Include="Events\PlayerJoinEvent.cs" />
     <Compile Include="Events\PlayerLeaveEvent.cs" />
-    <Compile Include="Events\PlayerProbeEvent.cs" />
+    <Compile Include="Tools\PlayerProbeEvent.cs" />
     <Compile Include="Events\PlayerReadyEvent.cs" />
-    <Compile Include="Events\PlayerSectorEvent.cs" />
+    <Compile Include="TransformSync\PlayerSectorEvent.cs" />
     <Compile Include="Events\PlayerStatesRequestEvent.cs" />
-    <Compile Include="Events\PlayerSuitEvent.cs" />
-    <Compile Include="Events\ServerTimeEvent.cs" />
+    <Compile Include="Animation\PlayerSuitEvent.cs" />
+    <Compile Include="TimeSync\ServerTimeEvent.cs" />
     <Compile Include="GeyserSync\GeyserEvent.cs" />
     <Compile Include="GeyserSync\GeyserManager.cs" />
     <Compile Include="GeyserSync\GeyserMessage.cs" />
     <Compile Include="GeyserSync\QSBGeyser.cs" />
-    <Compile Include="Messaging\AnimTriggerMessage.cs" />
-    <Compile Include="Messaging\PlayerDeathMessage.cs" />
+    <Compile Include="DeathSync\PlayerDeathMessage.cs" />
     <Compile Include="Messaging\PlayerLeaveMessage.cs" />
-    <Compile Include="Events\PlayerProbeLauncherEvent.cs" />
-    <Compile Include="Events\PlayerSignalscopeEvent.cs" />
-    <Compile Include="Events\PlayerTranslatorEvent.cs" />
+    <Compile Include="Tools\PlayerProbeLauncherEvent.cs" />
+    <Compile Include="Tools\PlayerSignalscopeEvent.cs" />
+    <Compile Include="Tools\PlayerTranslatorEvent.cs" />
     <Compile Include="Events\QSBEvent.cs" />
     <Compile Include="Messaging\PlayerJoinMessage.cs" />
     <Compile Include="Messaging\ToggleMessage.cs" />
@@ -159,15 +160,15 @@
     <Compile Include="Events\EventList.cs" />
     <Compile Include="Messaging\PlayerStateMessage.cs" />
     <Compile Include="Events\PlayerState.cs" />
-    <Compile Include="Events\Necronomicon.cs" />
+    <Compile Include="DeathSync\Necronomicon.cs" />
     <Compile Include="Utility\DebugLog.cs" />
     <Compile Include="Messaging\PlayerMessage.cs" />
     <Compile Include="Messaging\MessageHandler.cs" />
     <Compile Include="Messaging\MessageType.cs" />
     <Compile Include="PlayerInfo.cs" />
     <Compile Include="State.cs" />
-    <Compile Include="TimeSync\PreventShipDestruction.cs" />
-    <Compile Include="TimeSync\RespawnOnDeath.cs" />
+    <Compile Include="DeathSync\PreventShipDestruction.cs" />
+    <Compile Include="DeathSync\RespawnOnDeath.cs" />
     <Compile Include="PlayerRegistry.cs" />
     <Compile Include="TransformSync\PlayerCameraSync.cs" />
     <Compile Include="Utility\FlagsHelper.cs" />
@@ -179,9 +180,9 @@
     <Compile Include="TimeSync\PreserveTimeScale.cs" />
     <Compile Include="TransformSync\ShipTransformSync.cs" />
     <Compile Include="TransformSync\TransformSync.cs" />
-    <Compile Include="Messaging\SectorMessage.cs" />
+    <Compile Include="TransformSync\SectorMessage.cs" />
     <Compile Include="TransformSync\SectorSync.cs" />
-    <Compile Include="Messaging\ServerTimeMessage.cs" />
+    <Compile Include="TimeSync\ServerTimeMessage.cs" />
     <Compile Include="TimeSync\WakeUpSync.cs" />
     <Compile Include="QSBNetworkManager.cs" />
     <Compile Include="QSB.cs" />
